refactor(tcm): update TCM encounter model to use BaseModel and add eTag
metriport/metriport-internal#1040

Ref: ENG-479
