+++ packages/api/src/routes/medical/__tests__/tcm-encounter.test.ts
@@ -5,7 +5,7 @@ import { Op } from "sequelize";
 import { mockStartTransaction } from "../../../models/__tests__/transaction";
 import { makeTcmEncounterModel } from "../../../models/medical/__tests__/tcm-encounter";
 import { TcmEncounterModel } from "../../../models/medical/tcm-encounter";
-import { TcmEncounterCreate, TcmEncounterUpdate } from "../schemas/tcm-encounter";
+import { TcmEncounterUpdate } from "../schemas/tcm-encounter";
 
 jest.mock("../../../models/medical/tcm-encounter");
 
@@ -20,75 +20,20 @@ describe("TCM Encounter Routes", () => {
       params,
       query,
       cxId: faker.string.uuid(),
-    } as unknown as Request;
+    } as Request;
   };
 
   const mockResponse = () => {
     const res = {} as Response;
     res.status = jest.fn().mockReturnValue(res);
     res.json = jest.fn().mockReturnValue(res);
-    res.sendStatus = jest.fn().mockReturnValue(res);
     return res;
   };
 
   beforeEach(() => {
+    jest.clearAllMocks();
     jest.restoreAllMocks();
     mockStartTransaction();
-  });
-
-  describe("POST /dash-oss/medical/v1/tcm/encounter", () => {
-    const validCreatePayload: TcmEncounterCreate = {
-      patientId: faker.string.uuid(),
-      facilityName: faker.company.name(),
-      latestEvent: "Admitted",
-      class: "Inpatient",
-      admitTime: faker.date.recent().toISOString(),
-      dischargeTime: null,
-      clinicalInformation: { test: "data" },
-    };
-
-    it("creates a new TCM encounter", async () => {
-      const req = mockRequest(validCreatePayload);
-      const res = mockResponse();
-
-      const mockEncounter = makeTcmEncounterModel({
-        id: faker.string.uuid(),
-        cxId: req.cxId,
-        ...validCreatePayload,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        version: 1,
-      });
-
-      (TcmEncounterModel.create as jest.Mock).mockResolvedValueOnce(mockEncounter);
-
-      // TODO: Call the route handler
-      // await createTcmEncounter(req, res);
-
-      expect(TcmEncounterModel.create).toHaveBeenCalledWith({
-        cxId: req.cxId,
-        ...validCreatePayload,
-      });
-      expect(res.status).toHaveBeenCalledWith(httpStatus.CREATED);
-      expect(res.json).toHaveBeenCalledWith(mockEncounter);
-    });
-
-    it("validates required fields", async () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const req = mockRequest({});
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const res = mockResponse();
-
-      // TODO: Call the route handler
-      // await createTcmEncounter(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(httpStatus.BAD_REQUEST);
-      expect(res.json).toHaveBeenCalledWith(
-        expect.objectContaining({
-          message: expect.stringContaining("Required"),
-        })
-      );
-    });
   });
 
   describe("PUT /dash-oss/medical/v1/tcm/encounter/:id", () => {
@@ -99,9 +44,7 @@ describe("TCM Encounter Routes", () => {
 
     it("updates an existing TCM encounter", async () => {
       const encounterId = faker.string.uuid();
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const req = mockRequest(validUpdatePayload, { id: encounterId });
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const res = mockResponse();
 
       const mockEncounter = makeTcmEncounterModel({
@@ -116,7 +59,6 @@ describe("TCM Encounter Routes", () => {
         clinicalInformation: { test: "data" },
         createdAt: new Date(),
         updatedAt: new Date(),
-        version: 2,
       });
 
       (TcmEncounterModel.findByPk as jest.Mock).mockResolvedValueOnce(mockEncounter);
@@ -129,7 +71,6 @@ describe("TCM Encounter Routes", () => {
       expect(TcmEncounterModel.update).toHaveBeenCalledWith(
         {
           ...validUpdatePayload,
-          version: mockEncounter.version + 1,
         },
         {
           where: {
@@ -144,6 +85,7 @@ describe("TCM Encounter Routes", () => {
 
     it("returns 404 for non-existent encounter", async () => {
       const encounterId = faker.string.uuid();
+      // TODO: Remove these ESLint disable comments once the handler calls are implemented
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const req = mockRequest(validUpdatePayload, { id: encounterId });
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -166,9 +108,7 @@ describe("TCM Encounter Routes", () => {
 
   describe("GET /dash-oss/medical/v1/tcm/encounter", () => {
     it("returns paginated list of encounters", async () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const req = mockRequest(undefined, undefined, { count: "10" });
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const res = mockResponse();
 
       const mockEncounters = Array.from({ length: 10 }, () =>
@@ -184,7 +124,6 @@ describe("TCM Encounter Routes", () => {
           clinicalInformation: { test: "data" },
           createdAt: new Date(),
           updatedAt: new Date(),
-          version: 1,
         })
       );
 
@@ -241,8 +180,8 @@ describe("TCM Encounter Routes", () => {
     });
 
     it("applies default filter for admit time > 2020", async () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const req = mockRequest();
+
       // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const res = mockResponse();
